# [SPARK-19555][SQL] Improve the performance of StringUtils.escapeLikeRegex method